Convert timeout tests to mocha
