+++ mocha_test/timeout.js
@@ -1,5 +1,5 @@
 var async = require('../lib');
-var expect = require('chai').expect;
+var assert = require('assert');
 
 describe('timeout', function () {
 
@@ -17,9 +17,33 @@ describe('timeout', function () {
             }, 50)
         ],
         function(err, results) {
-            expect(err.message).to.equal('Callback function "asyncFn" timed out.');
-            expect(err.code).to.equal('ETIMEDOUT');
-            expect(results[0]).to.equal('I didn\'t time out');
+            assert(err.message === 'Callback function "asyncFn" timed out.');
+            assert(err.code === 'ETIMEDOUT');
+            assert(err.info === undefined);
+            assert(results[0] === 'I didn\'t time out');
+            done();
+        });
+    });
+
+    it('timeout with series and info', function (done) {
+        var info = { custom: 'info about callback' };
+        async.series([
+            async.timeout(function asyncFn(callback) {
+                setTimeout(function() {
+                    callback(null, 'I didn\'t time out');
+                }, 25);
+            }, 50),
+            async.timeout(function asyncFn(callback) {
+                setTimeout(function() {
+                    callback(null, 'I will time out');
+                }, 75);
+            }, 50, info)
+        ],
+        function(err, results) {
+            assert(err.message === 'Callback function "asyncFn" timed out.');
+            assert(err.code === 'ETIMEDOUT');
+            assert(err.info === info);
+            assert(results[0] === 'I didn\'t time out');
             done();
         });
     });
@@ -38,11 +62,11 @@ describe('timeout', function () {
             }, 50)
         ],
         function(err, results) {
-            expect(err.message).to.equal('Callback function "asyncFn" timed out.');
-            expect(err.code).to.equal('ETIMEDOUT');
-            expect(results[0]).to.equal('I didn\'t time out');
+            assert(err.message === 'Callback function "asyncFn" timed out.');
+            assert(err.code === 'ETIMEDOUT');
+            assert(err.info === undefined);
+            assert(results[0] === 'I didn\'t time out');
             done();
         });
     });
-
 });
